+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - "Create Variable" created a mess when used in some patterns (if, while, for statements). It now works as expected.
+- "Create Function" now infers enum types correctly in TS.
 
 ## [1.4.0] - 2020-05-21